Column attributes now use no m_ prefix
It was decided that it was a cleaner implementation to name the
ORM columns and ORM attributes differently.

This is cleaner without having to reference the m_ prefix all the
time, and the user should not have to know about this.

To avoid having m_ prefix on the attributes but still maintaining
prefix on column keys required an extra arg to be added to most
functions. Also fixed the bug where if an attribute was
previously called:
m_literaturesource_m_literaturesource_creator_m_\
literaturesource_creator_person_givenname

it would now be called:
m_literaturesource_literaturesource_creator_\
literaturesource_creator_person_givenname

(m_'s removed from within the name)

Question raised about whether we want the attribute to be called:
m_literaturesource_creator_person_givenname
but on attempting to modify this, an mapping error is called.