Reimplementation of schema logic and parsing data
Cross-code change to the way the schemas have been implemented and where
they are used. Each fields that will be output from a schema is uniquely
referenced so that the same schema can be used for validataion on
deserialization and also for producing json output.

Added measured_depth to the borehole level.

Using a custom parser with strict validation on the input query
parameters, as when using webargs, inputing a parameter that is not
included in the schema would not return an error. I have added the tests
for the strict.py parser file but these do not pass yet.