Do not return sections if epoch time not valid
When datetime query parameters were used to filter hydraulics,
The sections of the borehole would all be returned regardless
of the section epoch time. Now logic has been added to also
filter sections on epoch time if this does not overlap with
the hydraulic search criteria.

A test case was added to the populate_db_cases.py to test.