Addition of logic so that parents always returned if no children
There is a specific case where if hydraulic query parameters are
used and these filter out all the hydraulic children, then nothing
was returned. We thought the expection should be that if there
are no matching children, the parents should still be returned if
they are valid.

Validation on query parameters has also been updated in accordance
with the Ramsis core code so that these match.