Change to Literature Source & Creation Info
Exploring the usage of the WS with Postgresql, it was found that the
column names for flattened tables LiteratureSource etc. were too long
for the character limit, and these were truncated so that the db could
not be built.
The >63 char limit meant that the only sensible solution was to
reimplement the table structure as being nested tables.

Test files have also been moved to appropriately named folders.

Also some cleaning up of moved files was done.