[ENH] Creating POST route
A POST route is the best option for allowing user credentials to be
authorised when loading data to the db. The previous command line option
is still available as before (load_data.py) and both of these options
share the same functionality to merge data with existing data in the db.

Still to do: actually adding user credentials.