[ENH] Auto section start and end time enabled
If a section is uploaded, the default behaviour will be to update the
section start and end time based on the hydraulics given. If the
provided section start and end time are to be used, then the option
--auto_datetime_off should be used. I made this the default option to
avoid situations where the hydraulics and section times do not match,
which can lead to errors at the api query level.